+++ Source/Core/FontFace.cpp
@@ -26,16 +26,15 @@
  */
 
 #include "precompiled.h"
-#include "FontFace.h"
+#include "../../Include/Rocket/Core/FontFace.h"
 #include "FontFaceHandle.h"
 #include "../../Include/Rocket/Core/Log.h"
 
 namespace Rocket {
 namespace Core {
 
-FontFace::FontFace(FT_Face _face, Font::Style _style, Font::Weight _weight, bool _release_stream)
+FontFace::FontFace(Font::Style _style, Font::Weight _weight, bool _release_stream)
 {
-	face = _face;
 	style = _style;
 	weight = _weight;
 
@@ -50,8 +49,6 @@ FontFace::~FontFace()
 		for (size_t i = 0; i < handle_list.size(); ++i)
 			handle_list[i]->RemoveReference();
 	}
-
-	ReleaseFace();
 }
 
 // Returns the style of the font face.
@@ -64,96 +61,6 @@ Font::Style FontFace::GetStyle() const
 Font::Weight FontFace::GetWeight() const
 {
 	return weight;
-}
-
-// Returns a handle for positioning and rendering this face at the given size.
-FontFaceHandle* FontFace::GetHandle(const String& _raw_charset, int size)
-{
-	UnicodeRangeList charset;
-
-	HandleMap::iterator iterator = handles.find(size);
-	if (iterator != handles.end())
-	{
-		const HandleList& handles = (*iterator).second;
-
-		// Check all the handles if their charsets match the requested one exactly (ie, were specified by the same
-		// string).
-		String raw_charset(_raw_charset);
-		for (size_t i = 0; i < handles.size(); ++i)
-		{
-			if (handles[i]->GetRawCharset() == _raw_charset)
-			{
-				handles[i]->AddReference();
-				return handles[i];
-			}
-		}
-
-		// Check all the handles if their charsets contain the requested charset.
-		if (!UnicodeRange::BuildList(charset, raw_charset))
-		{
-			Log::Message(Log::LT_ERROR, "Invalid font charset '%s'.", _raw_charset.CString());
-			return NULL;
-		}
-
-		for (size_t i = 0; i < handles.size(); ++i)
-		{
-			bool range_contained = true;
-
-			const UnicodeRangeList& handle_charset = handles[i]->GetCharset();
-			for (size_t j = 0; j < charset.size() && range_contained; ++j)
-			{
-				if (!charset[j].IsContained(handle_charset))
-					range_contained = false;
-			}
-
-			if (range_contained)
-			{
-				handles[i]->AddReference();
-				return handles[i];
-			}
-		}
-	}
-
-	// See if this face has been released.
-	if (face == NULL)
-	{
-		Log::Message(Log::LT_WARNING, "Font face has been released, unable to generate new handle.");
-		return NULL;
-	}
-
-	// Construct and initialise the new handle.
-	FontFaceHandle* handle = new FontFaceHandle();
-	if (!handle->Initialise(face, _raw_charset, size))
-	{
-		handle->RemoveReference();
-		return NULL;
-	}
-
-	// Save the handle, and add a reference for the callee. The initial reference will be removed when the font face
-	// releases it.
-	if (iterator != handles.end())
-		(*iterator).second.push_back(handle);
-	else
-		handles[size] = HandleList(1, handle);
-
-	handle->AddReference();
-
-	return handle;
-}
-
-// Releases the face's FreeType face structure.
-void FontFace::ReleaseFace()
-{
-	if (face != NULL)
-	{
-		FT_Byte* face_memory = face->stream->base;
-		FT_Done_Face(face);
-
-		if (release_stream)
-			delete[] face_memory;
-
-		face = NULL;
-	}
 }
 
 }
