+++ Source/Core/FreeType/FontFace.h
@@ -1,82 +1,69 @@
-///*
-// * This source file is part of libRocket, the HTML/CSS Interface Middleware
-// *
-// * For the latest information, see http://www.librocket.com
-// *
-// * Copyright (c) 2008-2010 CodePoint Ltd, Shift Technology Ltd
-// *
-// * Permission is hereby granted, free of charge, to any person obtaining a copy
-// * of this software and associated documentation files (the "Software"), to deal
-// * in the Software without restriction, including without limitation the rights
-// * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// * copies of the Software, and to permit persons to whom the Software is
-// * furnished to do so, subject to the following conditions:
-// *
-// * The above copyright notice and this permission notice shall be included in
-// * all copies or substantial portions of the Software.
-// *
-// * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// * THE SOFTWARE.
-// *
-// */
+/*
+ * This source file is part of libRocket, the HTML/CSS Interface Middleware
+ *
+ * For the latest information, see http://www.librocket.com
+ *
+ * Copyright (c) 2008-2010 CodePoint Ltd, Shift Technology Ltd
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ */
 
-//#ifndef ROCKETCOREFONTFACE_H
-//#define ROCKETCOREFONTFACE_H
+#ifndef ROCKETCOREFREETYPEFONTFACE_H
+#define ROCKETCOREFREETYPEFONTFACE_H
 
-//#include "../../../Include/Rocket/Core/Font.h"
-//#include <ft2build.h>
-//#include FT_FREETYPE_H
+#include "../../../Include/Rocket/Core/FontFace.h"
+#include <ft2build.h>
+#include FT_FREETYPE_H
 
-//namespace Rocket {
-//namespace Core {
+namespace Rocket {
+namespace Core {
 
-//class FontFaceHandle;
+class FontFaceHandle;
 
-///**
-//	@author Peter Curry
-// */
+namespace FreeType {
+/**
+    @author Peter Curry
+ */
 
-//class FontFace
-//{
-//public:
-//	FontFace(FT_Face face, Font::Style style, Font::Weight weight, bool release_stream);
-//	~FontFace();
+class FontFace : public Rocket::Core::FontFace
+{
+public:
+    FontFace(FT_Face face, Font::Style style, Font::Weight weight, bool release_stream);
+    ~FontFace();
 
-//	/// Returns the style of the font face.
-//	/// @return The font face's style.
-//	Font::Style GetStyle() const;
-//	/// Returns the weight of the font face.
-//	/// @return The font face's weight.
-//	Font::Weight GetWeight() const;
+    /// Returns a handle for positioning and rendering this face at the given size.
+    /// @param[in] charset The set of characters in the handle, as a comma-separated list of unicode ranges.
+    /// @param[in] size The size of the desired handle, in points.
+    /// @return The shared font handle.
+    FontFaceHandle* GetHandle(const String& charset, int size);
 
-//	/// Returns a handle for positioning and rendering this face at the given size.
-//	/// @param[in] charset The set of characters in the handle, as a comma-separated list of unicode ranges.
-//	/// @param[in] size The size of the desired handle, in points.
-//	/// @return The shared font handle.
-//	FontFaceHandle* GetHandle(const String& charset, int size);
+    /// Releases the face's FreeType face structure. This will mean handles for new sizes cannot be constructed,
+    /// but existing ones can still be fetched.
+    void ReleaseFace();
 
-//	/// Releases the face's FreeType face structure. This will mean handles for new sizes cannot be constructed,
-//	/// but existing ones can still be fetched.
-//	void ReleaseFace();
+private:
+    FT_Face face;
+};
 
-//private:
-//	FT_Face face;
-//	Font::Style style;
-//	Font::Weight weight;
+}
+}
+}
 
-//	bool release_stream;
-
-//	typedef std::vector< FontFaceHandle* > HandleList;
-//	typedef std::map< int, HandleList > HandleMap;
-//	HandleMap handles;
-//};
-
-//}
-//}
-
-//#endif
+#endif
