+++ Source/Core/FreeType/FontFamily.h
@@ -57,21 +57,7 @@ public:
     /// @param[in] weight The weight of the new face.
     /// @param[in] release_stream True if the application must free the face's memory stream.
     /// @return True if the face was loaded successfully, false otherwise.
-    bool AddFace(FT_Face ft_face, Font::Style style, Font::Weight weight, bool release_stream);
-
-    /// Returns a handle to the most appropriate font in the family, at the correct size.
-    /// @param[in] charset The set of characters in the handle, as a comma-separated list of unicode ranges.
-    /// @param[in] style The style of the desired handle.
-    /// @param[in] weight The weight of the desired handle.
-    /// @param[in] size The size of desired handle, in points.
-    /// @return A valid handle if a matching (or closely matching) font face was found, NULL otherwise.
-    FontFaceHandle* GetFaceHandle(const String& charset, Font::Style style, Font::Weight weight, int size);
-
-private:
-    String name;
-
-    typedef std::vector< FontFace* > FontFaceList;
-    FontFaceList font_faces;
+    bool AddFace(void* ft_face, Font::Style style, Font::Weight weight, bool release_stream);
 };
 
 }
